# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Ardi - Autonomous Penetration Testing Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /scans:
    get:
      operationId: getScans
      tags: [scans]
      summary: List all scans
      responses:
        "200":
          description: Scan list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Scan"
    post:
      operationId: createScan
      tags: [scans]
      summary: Create a new scan
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScanInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Scan"
  /scans/{id}:
    get:
      operationId: getScan
      tags: [scans]
      summary: Get scan by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Scan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Scan"
  /scans/{id}/start:
    post:
      operationId: startScan
      tags: [scans]
      summary: Start a scan
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Scan started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Scan"
  /scans/{id}/stop:
    post:
      operationId: stopScan
      tags: [scans]
      summary: Stop a running scan
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Scan stopped
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Scan"
  /scans/{id}/findings:
    get:
      operationId: getScanFindings
      tags: [scans]
      summary: Get findings for a specific scan
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Findings for scan
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Finding"

  # Findings
...
  /reports:
    get:
      operationId: getReports
      tags: [reports]
      summary: List all reports
      responses:
        "200":
          description: Reports list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Report"
    post:
      operationId: createReport
      tags: [reports]
      summary: Generate a new report
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReportInput"
      responses:
        "201":
          description: Report created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Report"
  /reports/{id}:
    get:
      operationId: getReport
      tags: [reports]
      summary: Get report by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Report"

  # Compliance
...
components:
  schemas:
...
    Scan:
      type: object
      required: [id, name, type, status, assetId, assetName, createdAt]
      properties:
        id:
          type: number
        name:
          type: string
        type:
          type: string
          enum: [web_app, network, cloud, api, full_stack]
        status:
          type: string
          enum: [pending, running, completed, failed, stopped]
        assetId:
          type: number
        assetName:
          type: string
        progress:
          type: number
        startedAt:
          type: ["string", "null"]
        completedAt:
          type: ["string", "null"]
        findingsCount:
          type: number
        criticalCount:
          type: number
        highCount:
          type: number
        duration:
          type: ["number", "null"]
        createdAt:
          type: string
    ScanInput:
      type: object
      required: [name, type, assetId]
      properties:
        name:
          type: string
        type:
          type: string
          enum: [web_app, network, cloud, api, full_stack]
        assetId:
          type: number
        config:
          type: object
          properties:
            depth:
              type: string
              enum: [shallow, standard, deep]
            authenticated:
              type: boolean
            modules:
              type: array
              items:
                type: string

    # Findings
    Finding:
      type: object
      required: [id, title, severity, status, category, assetId, assetName, scanId, cve, mitreId, mitreTactic, mitreTechnique, description, remediation, createdAt]
      properties:
        id:
          type: number
        title:
          type: string
        severity:
          type: string
          enum: [critical, high, medium, low, info]
        status:
          type: string
          enum: [open, in_progress, resolved, accepted_risk, false_positive]
        category:
          type: string
          # Web-application categories, plus the infrastructure categories the
          # nmap and nuclei engines actually emit (network, tls, dns, exposure,
          # web, file, other). The engine is the source of truth here: a
          # category it can produce but the spec rejects makes the whole
          # findings endpoint fail Zod validation and return 500.
          enum: [injection, broken_auth, xss, idor, misconfiguration, exposed_data, privilege_escalation, lateral_movement, persistence, exfiltration, network, web, tls, dns, exposure, file, other]
        assetId:
          type: number
        assetName:
          type: string
        scanId:
          type: number
        cve:
          type: ["string", "null"]
        cvss:
          type: ["number", "null"]
        mitreId:
          type: string
        mitreTactic:
          type: string
        mitreTechnique:
          type: string
        description:
          type: string
        remediation:
          type: string
        evidence:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        resolvedAt:
          type: ["string", "null"]
        createdAt:
          type: string
...
    Report:
      type: object
      required: [id, title, type, status, format, createdAt]
      properties:
        id:
          type: number
        title:
          type: string
        type:
          type: string
          enum: [executive, technical, compliance, pentest]
        status:
          type: string
          enum: [generating, ready, failed]
        format:
          type: string
          enum: [pdf, html, json]
        scanId:
          type: ["number", "null"]
        assetId:
          type: ["number", "null"]
        summary:
          type: ["string", "null"]
        downloadUrl:
          type: ["string", "null"]
        createdAt:
          type: string
    ReportInput:
      type: object
      required: [title, type, format]
      properties:
        title:
          type: string
        type:
          type: string
          enum: [executive, technical, compliance, pentest]
        format:
          type: string
          enum: [pdf, html, json]
        scanId:
          type: number
        assetId:
          type: number
        dateRange:
          type: object
          properties:
            from:
              type: string
            to:
              type: string

    # Compliance
...